+++ lib/boruta/adapters/ecto/schemas/token.ex
@@ -24,6 +24,7 @@ defmodule Boruta.Ecto.Token do
           nonce: String.t(),
           c_nonce: String.t(),
           scope: String.t(),
+          resource: String.t() | nil,
           redirect_uri: String.t(),
           expires_at: integer(),
           client: Client.t(),
@@ -69,6 +70,7 @@ defmodule Boruta.Ecto.Token do
     field(:nonce, :string)
     field(:c_nonce, :string)
     field(:scope, :string, default: "")
+    field(:resource, :string)
     field(:redirect_uri, :string)
     field(:expires_at, :integer)
     field(:revoked_at, :utc_datetime_usec)
@@ -103,6 +105,7 @@ defmodule Boruta.Ecto.Token do
       :state,
       :nonce,
       :scope,
+      :resource,
       :access_token_ttl,
       :previous_code,
       :authorization_details,
@@ -129,6 +132,7 @@ defmodule Boruta.Ecto.Token do
       :state,
       :nonce,
       :scope,
+      :resource,
       :previous_token,
       :previous_code,
       :authorization_details,
@@ -153,6 +157,7 @@ defmodule Boruta.Ecto.Token do
       :state,
       :nonce,
       :scope,
+      :resource,
       :access_token_ttl,
       :previous_code,
       :authorization_details,
@@ -179,6 +184,7 @@ defmodule Boruta.Ecto.Token do
       :state,
       :nonce,
       :scope,
+      :resource,
       :previous_token,
       :previous_code,
       :authorization_details,
@@ -204,6 +210,7 @@ defmodule Boruta.Ecto.Token do
       :state,
       :nonce,
       :scope,
+      :resource,
       :authorization_details,
       :redirect_uri,
       :agent_token
@@ -226,6 +233,7 @@ defmodule Boruta.Ecto.Token do
       :state,
       :nonce,
       :scope,
+      :resource,
       :code_challenge,
       :code_challenge_method,
       :authorization_details,
@@ -259,6 +267,7 @@ defmodule Boruta.Ecto.Token do
       :state,
       :nonce,
       :scope,
+      :resource,
       :authorization_details,
       :presentation_definition
     ])
@@ -281,6 +290,7 @@ defmodule Boruta.Ecto.Token do
       :state,
       :nonce,
       :scope,
+      :resource,
       :code_challenge,
       :code_challenge_method,
       :authorization_details,