+++ lib/boruta/oauth/json/schema.ex
@@ -21,7 +21,8 @@ defmodule Boruta.Oauth.Json.Schema do
           },
           "required" => ["type", "value"]
         },
-        "scope" => %{"type" => "string"}
+        "scope" => %{"type" => "string"},
+        "resource" => %{"type" => "string"}
       },
       "required" => ["grant_type", "client_id", "client_authentication"]
     }
@@ -46,6 +47,7 @@ defmodule Boruta.Oauth.Json.Schema do
           "required" => ["type", "value"]
         },
         "scope" => %{"type" => "string"},
+        "resource" => %{"type" => "string"},
         "bind_data" => %{"type" => "string"},
         "bind_configuration" => %{"type" => "string"}
       },
@@ -73,7 +75,8 @@ defmodule Boruta.Oauth.Json.Schema do
         },
         "username" => %{"type" => "string"},
         "password" => %{"type" => "string"},
-        "scope" => %{"type" => "string"}
+        "scope" => %{"type" => "string"},
+        "resource" => %{"type" => "string"}
       },
       "required" => ["grant_type", "client_id", "username", "password"]
     }
@@ -98,6 +101,7 @@ defmodule Boruta.Oauth.Json.Schema do
         },
         "code" => %{"type" => "string"},
         "redirect_uri" => %{"type" => "string"},
+        "resource" => %{"type" => "string"},
         "code_verifier" => %{"type" => "string"}
       },
       "required" => ["grant_type", "code", "client_id"]
@@ -123,6 +127,7 @@ defmodule Boruta.Oauth.Json.Schema do
         },
         "code" => %{"type" => "string"},
         "redirect_uri" => %{"type" => "string"},
+        "resource" => %{"type" => "string"},
         "code_verifier" => %{"type" => "string"},
         "bind_data" => %{"type" => "string"},
         "bind_configuration" => %{"type" => "string"}
@@ -166,6 +171,7 @@ defmodule Boruta.Oauth.Json.Schema do
         "state" => %{"type" => "string"},
         "nonce" => %{"type" => "string"},
         "redirect_uri" => %{"type" => "string"},
+        "resource" => %{"type" => "string"},
         "prompt" => %{"type" => "string"}
       },
       "required" => ["response_type", "client_id", "redirect_uri"]
@@ -184,7 +190,8 @@ defmodule Boruta.Oauth.Json.Schema do
         },
         "state" => %{"type" => "string"},
         "nonce" => %{"type" => "string"},
-        "redirect_uri" => %{"type" => "string"}
+        "redirect_uri" => %{"type" => "string"},
+        "resource" => %{"type" => "string"}
       },
       "required" => ["response_type", "client_id", "redirect_uri"]
     }
@@ -201,7 +208,8 @@ defmodule Boruta.Oauth.Json.Schema do
         "state" => %{"type" => "string"},
         "client_metadata" => %{"type" => "string"},
         "nonce" => %{"type" => "string"},
-        "redirect_uri" => %{"type" => "string"}
+        "redirect_uri" => %{"type" => "string"},
+        "resource" => %{"type" => "string"}
       },
       "required" => ["response_type", "client_id", "redirect_uri"]
     }
@@ -214,7 +222,8 @@ defmodule Boruta.Oauth.Json.Schema do
       "properties" => %{
         "grant_type" => %{"type" => "string", "pattern" => "refresh_token"},
         "refresh_token" => %{"type" => "string"},
-        "scope" => %{"type" => "string"}
+        "scope" => %{"type" => "string"},
+        "resource" => %{"type" => "string"}
       },
       "required" => ["grant_type", "refresh_token"]
     }
@@ -233,6 +242,7 @@ defmodule Boruta.Oauth.Json.Schema do
         "state" => %{"type" => "string"},
         "nonce" => %{"type" => "string"},
         "redirect_uri" => %{"type" => "string"},
+        "resource" => %{"type" => "string"},
         "prompt" => %{"type" => "string"},
         "code_challenge" => %{"type" => "string"},
         "code_challenge_method" => %{
@@ -257,6 +267,7 @@ defmodule Boruta.Oauth.Json.Schema do
         "state" => %{"type" => "string"},
         "nonce" => %{"type" => "string"},
         "redirect_uri" => %{"type" => "string"},
+        "resource" => %{"type" => "string"},
         "prompt" => %{"type" => "string"},
         "code_challenge" => %{"type" => "string"},
         "code_challenge_method" => %{