+++ lib/boruta/oauth/request/base.ex
@@ -36,6 +36,7 @@ defmodule Boruta.Oauth.Request.Base do
        client_id: params["client_id"],
        client_authentication: client_authentication_from_params(params),
        scope: params["scope"],
+       resource: params["resource"],
        dpop: params["dpop"]
      }}
   end
@@ -46,6 +47,7 @@ defmodule Boruta.Oauth.Request.Base do
        client_id: params["client_id"],
        client_authentication: client_authentication_from_params(params),
        scope: params["scope"],
+       resource: params["resource"],
        dpop: params["dpop"],
        bind_data: params["bind_data"],
        bind_configuration: params["bind_configuration"]
@@ -59,7 +61,8 @@ defmodule Boruta.Oauth.Request.Base do
        client_authentication: client_authentication_from_params(params),
        username: params["username"],
        password: params["password"],
-       scope: params["scope"]
+       scope: params["scope"],
+       resource: params["resource"]
      }}
   end
 
@@ -70,6 +73,7 @@ defmodule Boruta.Oauth.Request.Base do
        client_authentication: client_authentication_from_params(params),
        code: params["code"],
        redirect_uri: params["redirect_uri"],
+       resource: params["resource"],
        code_verifier: params["code_verifier"],
        dpop: params["dpop"]
      }}
@@ -82,6 +86,7 @@ defmodule Boruta.Oauth.Request.Base do
        client_authentication: client_authentication_from_params(params),
        code: params["code"],
        redirect_uri: params["redirect_uri"],
+       resource: params["resource"],
        code_verifier: params["code_verifier"],
        dpop: params["dpop"],
        bind_data: params["bind_data"],
@@ -121,7 +126,8 @@ defmodule Boruta.Oauth.Request.Base do
        client_id: params["client_id"],
        client_authentication: client_authentication_from_params(params),
        refresh_token: params["refresh_token"],
-       scope: params["scope"]
+       scope: params["scope"],
+       resource: params["resource"]
      }}
   end
 
@@ -157,7 +163,8 @@ defmodule Boruta.Oauth.Request.Base do
       state: params["state"],
       code_challenge: params["code_challenge"],
       code_challenge_method: params["code_challenge_method"],
-      scope: params["scope"]
+      scope: params["scope"],
+      resource: params["resource"]
     }
 
     {:ok, request}
@@ -174,6 +181,7 @@ defmodule Boruta.Oauth.Request.Base do
       code_challenge: params["code_challenge"],
       code_challenge_method: params["code_challenge_method"],
       scope: params["scope"],
+      resource: params["resource"],
       response_mode: params["response_mode"]
     }
 
@@ -211,6 +219,7 @@ defmodule Boruta.Oauth.Request.Base do
           response_mode: params["response_mode"],
           response_types: response_types,
           scope: params["scope"],
+          resource: params["resource"],
           state: params["state"]
         }
 
@@ -233,6 +242,7 @@ defmodule Boruta.Oauth.Request.Base do
            response_types: response_types,
            response_mode: params["response_mode"],
            scope: params["scope"],
+           resource: params["resource"],
            state: params["state"]
          }}
     end