+++ priv/boruta/migrations/20260430120100_authorization_request_resource.ex
@@ -1,0 +1,13 @@
+defmodule Boruta.Migrations.AuthorizationRequestResource do
+  @moduledoc false
+
+  defmacro __using__(_args) do
+    quote do
+      def change do
+        alter table(:authorization_requests) do
+          add :resource, :string
+        end
+      end
+    end
+  end
+end